Move encryptoin tools to their own file
Just to simplify mod.rs but we probably should move all of this into
kernel interface so we can unit test it nicely.
